+++ packages/stream_chat_flutter/lib/src/gallery/gallery_footer.dart
@@ -217,8 +217,8 @@ class _StreamGalleryFooterState extends State<StreamGalleryFooter> {
                           child: AspectRatio(
                             aspectRatio: 1,
                             child: StreamVideoThumbnailImage(
-                              video: (attachment.file?.path ??
-                                  attachment.assetUrl)!,
+                              video:
+                                  attachment.file?.path ?? attachment.assetUrl,
                             ),
                           ),
                         ),