+++ packages/stream_chat_flutter/lib/src/message_widget/message_widget.dart
@@ -8,6 +8,7 @@ import 'package:stream_chat_flutter/src/context_menu_items/context_menu_reaction
 import 'package:stream_chat_flutter/src/context_menu_items/stream_chat_context_menu_item.dart';
 import 'package:stream_chat_flutter/src/dialogs/dialogs.dart';
 import 'package:stream_chat_flutter/src/message_actions_modal/message_actions_modal.dart';
+import 'package:stream_chat_flutter/src/message_widget/bottom_row.dart';
 import 'package:stream_chat_flutter/src/message_widget/message_widget_content.dart';
 import 'package:stream_chat_flutter/src/message_widget/reactions/message_reactions_modal.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
@@ -78,8 +79,15 @@ class StreamMessageWidget extends StatefulWidget {
     this.userAvatarBuilder,
     this.editMessageInputBuilder,
     this.textBuilder,
-    this.bottomRowBuilder,
-    this.deletedBottomRowBuilder,
+    @Deprecated('''
+    Use [bottomRowBuilderWithDefaultWidget] instead.
+    Will be removed in the next major version.
+    ''') this.bottomRowBuilder,
+    this.bottomRowBuilderWithDefaultWidget,
+    @Deprecated('''
+    Use [bottomRowBuilderWithDefaultWidget] instead.
+    Will be removed in the next major version.
+    ''') this.deletedBottomRowBuilder,
     this.customAttachmentBuilders,
     this.padding,
     this.textPadding = const EdgeInsets.symmetric(
@@ -91,11 +99,18 @@ class StreamMessageWidget extends StatefulWidget {
     this.onQuotedMessageTap,
     this.customActions = const [],
     this.onAttachmentTap,
-    this.usernameBuilder,
+    @Deprecated('''
+    Use [bottomRowBuilderWithDefaultWidget] instead.
+    Will be removed in the next major version.
+    ''') this.usernameBuilder,
     this.imageAttachmentThumbnailSize = const Size(400, 400),
     this.imageAttachmentThumbnailResizeType = 'clip',
     this.imageAttachmentThumbnailCropType = 'center',
-  }) : attachmentBuilders = {
+  })  : assert(
+          bottomRowBuilder == null || bottomRowBuilderWithDefaultWidget == null,
+          'You can only use one of the two bottom row builders',
+        ),
+        attachmentBuilders = {
           'image': (context, message, attachments) {
             final border = RoundedRectangleBorder(
               side: attachmentBorderSide ??
@@ -308,7 +323,13 @@ class StreamMessageWidget extends StatefulWidget {
   /// {@template bottomRowBuilder}
   /// Widget builder for building a bottom row below the message
   /// {@endtemplate}
-  final Widget Function(BuildContext, Message)? bottomRowBuilder;
+  final BottomRowBuilder? bottomRowBuilder;
+
+  /// {@template bottomRowBuilderWithDefaultWidget}
+  /// Widget builder for building a bottom row below the message.
+  /// Also contains the default bottom row widget.
+  /// {@endtemplate}
+  final BottomRowBuilderWithDefaultWidget? bottomRowBuilderWithDefaultWidget;
 
   /// {@template deletedBottomRowBuilder}
   /// Widget builder for building a bottom row below a deleted message
@@ -576,6 +597,26 @@ class StreamMessageWidget extends StatefulWidget {
     String? imageAttachmentThumbnailResizeType,
     String? imageAttachmentThumbnailCropType,
   }) {
+    assert(
+      bottomRowBuilder == null || bottomRowBuilderWithDefaultWidget == null,
+      'You can only use one of the two bottom row builders',
+    );
+
+    var _bottomRowBuilderWithDefaultWidget = bottomRowBuilderWithDefaultWidget;
+
+    _bottomRowBuilderWithDefaultWidget ??= (context, message, defaultWidget) {
+      final _bottomRowBuilder = bottomRowBuilder ?? this.bottomRowBuilder;
+      if (_bottomRowBuilder != null) {
+        return _bottomRowBuilder(context, message);
+      }
+
+      return defaultWidget.copyWith(
+        usernameBuilder: usernameBuilder ?? this.usernameBuilder,
+        deletedBottomRowBuilder:
+            deletedBottomRowBuilder ?? this.deletedBottomRowBuilder,
+      );
+    };
+
     return StreamMessageWidget(
       key: key ?? this.key,
       isDm: isDm ?? this.isDm,
@@ -586,10 +627,7 @@ class StreamMessageWidget extends StatefulWidget {
           editMessageInputBuilder ?? this.editMessageInputBuilder,
       botBuilder: botBuilder ?? this.botBuilder,
       textBuilder: textBuilder ?? this.textBuilder,
-      usernameBuilder: usernameBuilder ?? this.usernameBuilder,
-      bottomRowBuilder: bottomRowBuilder ?? this.bottomRowBuilder,
-      deletedBottomRowBuilder:
-          deletedBottomRowBuilder ?? this.deletedBottomRowBuilder,
+      bottomRowBuilderWithDefaultWidget: _bottomRowBuilderWithDefaultWidget,
       onMessageActions: onMessageActions ?? this.onMessageActions,
       message: message ?? this.message,
       messageTheme: messageTheme ?? this.messageTheme,
@@ -841,7 +879,6 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
                     showUserAvatar: widget.showUserAvatar,
                     streamChat: _streamChat,
                     translateUserAvatar: widget.translateUserAvatar,
-                    deletedBottomRowBuilder: widget.deletedBottomRowBuilder,
                     shape: widget.shape,
                     borderSide: widget.borderSide,
                     borderRadiusGeometry: widget.borderRadiusGeometry,
@@ -849,10 +886,7 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
                     onLinkTap: widget.onLinkTap,
                     onMentionTap: widget.onMentionTap,
                     onQuotedMessageTap: widget.onQuotedMessageTap,
-                    bottomRowBuilder: widget.bottomRowBuilder,
                     onUserAvatarTap: widget.onUserAvatarTap,
-                    userAvatarBuilder: widget.userAvatarBuilder,
-                    usernameBuilder: widget.usernameBuilder,
                   ),
                 ),
               ),