+++ packages/stream_chat_flutter/lib/src/misc/option_list_tile.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:stream_chat_flutter/src/theme/stream_chat_theme.dart';
 
 /// {@template streamOptionListTile}
@@ -54,7 +55,7 @@ class StreamOptionListTile extends StatelessWidget {
         Material(
           color: tileColor ?? chatThemeData.colorTheme.barsBg,
           child: SizedBox(
-            height: 63,
+            height: 40.h,
             child: InkWell(
               onTap: onTap,
               child: Row(